# backend/api-spec.yaml
openapi: 3.0.3
info:
  title: Test Case Manager API
  description: RESTful API for Test Case Management System with TestLink integration
  version: 1.0.0
  contact:
    name: Test Case Manager
    url: http://192.168.4.121:3000
servers:
  - url: http://192.168.4.121:3001/api
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /testcases:
    get:
      summary: List test cases
      tags: [Test Cases]
      parameters:
        - name: projectId
          in: query
          description: Filter by project ID
          schema:
            type: integer
        - name: test_suite_id
          in: query
          description: Filter by test suite ID
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, passed, failed, blocked, not_run]
        - name: priority
          in: query
          description: Filter by priority
          schema:
            type: string
            enum: [low, medium, high]
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Test cases retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestCase'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new test case
      tags: [Test Cases]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseInput'
      responses:
        '201':
          description: Test case created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /testcases/{id}:
    get:
      summary: Get test case by ID
      tags: [Test Cases]
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      responses:
        '200':
          description: Test case retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestCase'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update test case
      tags: [Test Cases]
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseInput'
      responses:
        '200':
          description: Test case updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete test case
      tags: [Test Cases]
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      responses:
        '200':
          description: Test case deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    TestCaseId:
      name: id
      in: path
      required: true
      description: Test Case ID
      schema:
        type: integer
  schemas:
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
          example: 1279
        test_suite_id:
          type: integer
          example: 105
        title:
          type: string
          example: "(D)Ad-WiFi-Client-Iso(RC)WiFi_AP(RE)(All Venues) Enable isolation in AP"
        description:
          type: string
        prerequisites:
          type: string
        execution_type:
          type: integer
          enum: [1, 2]
          description: "1=Manual, 2=Automated"
        external_id:
          type: string
          nullable: true
        internal_id:
          type: string
          nullable: true
        version:
          type: integer
        priority:
          type: integer
          enum: [1, 2, 3]
          description: "1=Low, 2=Medium, 3=High"
        importance:
          type: integer
          enum: [1, 2, 3]
        is_open:
          type: boolean
        active:
          type: boolean
        status:
          type: integer
          enum: [1, 2, 3, 4, 5]
          description: "1=Not Run, 2=Passed, 3=Failed, 4=Blocked, 5=Pending"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        test_suite_name:
          type: string
          example: "Network Control -> Client Isolation"
        project_name:
          type: string
          example: "Phase 3"
    TestCaseInput:
      type: object
      required: [title, test_steps, expected_result]
      properties:
        project_id:
          type: integer
        test_suite_id:
          type: integer
          nullable: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        prerequisites:
          type: string
        test_steps:
          type: string
        expected_result:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        execution_type:
          type: integer
          enum: [1, 2]
          default: 1
        importance:
          type: integer
          enum: [1, 2, 3]
          default: 2
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...